FIX: Json dump error when function tools
